# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "VistoriaPro API",
    "version": "1.0.0",
    "description": "Documentação da API VistoriaPro"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor local (desenvolvimento)"
    }
  ],
  "components": {
    "schemas": {
...
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "empresa_id": {
            "type": "integer"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "papel": {
            "type": "string",
            "enum": [
              "admin",
              "vistoriador",
              "cliente"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "empresa_id",
          "nome",
          "email"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Use 'Bearer <token>' to authenticate. Token obtained from /usuarios/auth/login."
      }
    }
  },
  "security": [],
  "paths": {
...
    "/usuarios": {
      "get": {
        "tags": [
          "Usuários"
        ],
        "summary": "Listar usuários",
        "description": "Retorna a lista de todos os usuários cadastrados.",
        "responses": {
          "200": {
            "description": "Lista de usuários retornada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Usuario"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno no servidor"
          }
        }
      }
    },
    "/usuarios/auth/login": {
      "post": {
        "tags": ["Usuários"],
        "summary": "Autenticar usuário (obter JWT)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Autenticação bem-sucedida. Retorna objeto com token JWT.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Requisição inválida." },
          "401": { "description": "Credenciais inválidas." }
        }
      }
    },
...
  },
...
}